+++ config/routes.rb
@@ -1,4 +1,4 @@
 Ckeditor::Engine.routes.draw do
-  resources :pictures, only: [:index, :create, :destroy]
-  resources :attachment_files, only: [:index, :create, :destroy]
+  resources :pictures, only: %i[index create destroy]
+  resources :attachment_files, only: %i[index create destroy]
 end