+++ lib/ckeditor/helpers/controllers.rb
@@ -20,7 +20,7 @@ module Ckeditor
       end
 
       def ckeditor_filebrowser_scope(options = {})
-        { order: [:id, :desc] }.merge!(options)
+        { order: %i[id desc] }.merge!(options)
       end
     end
   end