+++ lib/generators/ckeditor/templates/active_record/carrierwave/ckeditor/asset.rb
@@ -1,4 +1,6 @@
-class Ckeditor::Asset < ActiveRecord::Base
+# frozen_string_literal: true
+
+class Ckeditor::Asset < ApplicationRecord
   include Ckeditor::Orm::ActiveRecord::AssetBase
 
   delegate :url, :current_path, :content_type, to: :data