+++ test/models/picture_test.rb
@@ -8,12 +8,15 @@ class PictureTest < ActiveSupport::TestCase
   test "Set file content_type and size" do
     @picture = create_picture
 
-    assert_equal "image/png", @picture.data_content_type
-    assert_equal "rails.png", @picture.data_file_name
-    assert_equal 6646, @picture.data_file_size
+    assert_equal 'rails.png', @picture.data_file_name
 
-    if CKEDITOR_BACKEND == :dragonfly
+    case CKEDITOR_BACKEND
+    when :dragonfly
       assert @picture.url_thumb.include?('thumb_rails')
+    when :active_storage
+      assert @picture.url_thumb =~ /\/representations\/.*\/rails.png/
+    when :shrine
+      assert @picture.url_thumb =~ /\S{32}\.png/
     else
       assert @picture.url_thumb.include?('thumb_rails.png')
     end