+++ test/test_helper.rb
@@ -1,15 +1,14 @@
 # Configure Rails Envinronment
-ENV["RAILS_ENV"] = "test"
-CKEDITOR_ORM = (ENV["CKEDITOR_ORM"] || :active_record).to_sym
-CKEDITOR_BACKEND = (ENV["CKEDITOR_BACKEND"] || :paperclip).to_sym
+ENV['RAILS_ENV'] = 'test'
+CKEDITOR_ORM = (ENV['CKEDITOR_ORM'] || :active_record).to_sym
+CKEDITOR_BACKEND = (ENV['CKEDITOR_BACKEND'] || :carrierwave).to_sym
 
-$:.unshift File.dirname(__FILE__)
+$LOAD_PATH.unshift File.dirname(__FILE__)
 puts "\n==> Ckeditor.orm = #{CKEDITOR_ORM.inspect}. CKEDITOR_ORM = (active_record|mongoid)"
 puts "\n==> Ckeditor.backend = #{CKEDITOR_BACKEND.inspect}. CKEDITOR_BACKEND = (paperclip|carrierwave|refile|dragonfly)"
 
-require File.expand_path("../dummy/config/environment.rb", __FILE__)
-require "rails/test_help"
-require 'redgreen'
+require File.expand_path('dummy/config/environment.rb', __dir__)
+require 'rails/test_help'
 
 Rails.backtrace_cleaner.remove_silencers!
 
@@ -22,7 +21,7 @@ Capybara.default_selector = :css
 require "orm/#{CKEDITOR_ORM}"
 
 # Load support files
-$:.unshift File.expand_path('../support', __FILE__)
+$LOAD_PATH.unshift File.expand_path('support', __dir__)
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 # For generators
@@ -31,6 +30,10 @@ require "generators/ckeditor/install_generator"
 
 # Run template migration for the selected backend
 if CKEDITOR_ORM == :active_record
-  require "generators/ckeditor/templates/active_record/#{CKEDITOR_BACKEND}/migration.rb"
+  require "generators/ckeditor/templates/active_record/#{CKEDITOR_BACKEND}/create_ckeditor_assets.rb"
   CreateCkeditorAssets.new.up
 end
+
+# For create variants in Active Storage
+require 'active_job'
+ActiveJob::Base.queue_adapter = :test