+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Ckeditor changelog
 
+## Ckeditor 5.2.0
 
-## NEXT
+  * Upgrade CKEditor to 5 (44.1.0) version
+  * Bump ruby version to 3.2.6
+
+## Ckeditor 5.1.3
+
+  * Rename app/assets/config/manifest.js to app/assets/config/ckeditor_manifest.js (issue #938)
+  * Set config.versionCheck = false; by default to hide security notification (issue #933)
+
+## Ckeditor 5.1.2
+
+  * Upgrade CKEditor to 4.22.1 version
+  * Bump ruby version to 2.7.6
+  * Temporarily turn off paperclip gem support
 
 ## Ckeditor 5.1.1 (12-1-2021)
 
@@ -16,6 +29,8 @@
 ## Ckeditor 5.0.0
 
   * CDN version only (breaking changes)
+
+    Previously you could use a bundled version of the editor by adding `//= require ckeditor/init` to your application.js. From 5.0 this is no longer available.
   * Upgrade rails to 5.2.x
   * Remove refile support due no activity since 2015
   * Remove auto-detect content-type (`extract_content_type`) and image dimensions (`extract_dimensions`)